PC/requirements.txt
Package                   Version
------------------------- ------------
altgraph                  0.17.5
bidict                    0.23.1
blinker                   1.9.0
certifi                   2026.2.25
charset-normalizer        3.4.4
click                     8.3.1
colorama                  0.4.6
contourpy                 1.3.2
cycler                    0.12.1
Flask                     3.1.3
Flask-SocketIO            5.6.1
fonttools                 4.61.1
h11                       0.16.0
idna                      3.11
itsdangerous              2.2.0
Jinja2                    3.1.6
joblib                    1.5.3
kiwisolver                1.4.9
MarkupSafe                3.0.3
mediapipe                 0.10.9
MouseInfo                 0.1.3
packaging                 26.0
pandas                    2.3.3
pefile                    2024.8.26
pillow                    12.1.1
pip                       26.0.1
protobuf                  3.20.3
PyAutoGUI                 0.9.54
PyGetWindow               0.0.9
pyinstaller               6.19.0
pyinstaller-hooks-contrib 2026.2
PyMsgBox                  2.0.1
pynput                    1.8.1
pyparsing                 3.3.2
pyperclip                 1.11.0
PyQt5                     5.15.11
PyQt5-Qt5                 5.15.2
PyQt5_sip                 12.18.0
PyRect                    0.2.0
PyScreeze                 1.0.1
python-engineio           4.13.1
python-socketio           5.16.1
pytweening                1.2.0
pytz                      2025.2
pywin32-ctypes            0.2.3
requests                  2.32.5
schedule                  1.2.2
scikit-learn              1.7.2
scipy                     1.15.3
seaborn                   0.13.2
setuptools                82.0.0
simple-websocket          1.1.0
six                       1.17.0
threadpoolctl             3.6.0
torch                     2.1.2+cu118
torchaudio                2.1.2+cu118
torchvision               0.16.2+cu118
tqdm                      4.67.3
urllib3                   2.6.3
websocket-client          1.9.0
Werkzeug                  3.1.6
wheel                     0.46.3
wsproto                   1.3.2
